+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 #### Unreleased (dev branch)
+- Round robin and swiss now uses tiebreaker to rank all teams before transferring teams to the next stage.
 - Matches are now displayed slightly bigger
 - Matches now has an identifier and pending matches will display which matches they depend on. E.g. "Winner of X" vs "Loser of Y"